Find mount point using canonical path
When copying Lustre stripe information, the destination directory is checked to ensure that it resides on a Lustre filesystem. However, the original check failed if the destination path contained a symlink.

For example, suppose the destination is "/flush/dest", where "/flush" is a symlink to the mount point "/lustre/flush". The original code walks up the specified destination path until it finds a mount point, which in this case is "/". However, the modified code walks up the canonical path "/lustre/flush/dest", and the mount point is identified correctly as "/lustre/flush".